Windows Form App
Beginning of a windows form app for a RN which ask for some general answers pertaining to the user. If the user clicks yes the information is correct when the message box appears, then the user can move on to the application, but if they click no, then it close out the beginning stage of the application.